+++ src/infra/validators/email-validator-adapter.ts
@@ -1,4 +1,4 @@
-import { EmailValidator } from '../../../presentation/protocols/email-validator'
+import { EmailValidator } from '../../validation/protocols/email-validator'
 import validator from 'validator'
 
 export class EmailValidatorAdapter implements EmailValidator {
