+++ src/validation/validators/email-validation.ts
@@ -1,6 +1,5 @@
-import { InvalidParamError } from '../../errors'
-import { EmailValidator } from '../../protocols'
-import { Validation } from '../../protocols/validation'
+import { InvalidParamError } from '../../presentation/errors'
+import { EmailValidator, Validation } from '../../presentation/protocols'
 
 export class EmailValidation implements Validation {
   constructor (
