+++ src/validation/validators/index.ts
@@ -2,3 +2,4 @@ export * from './compare-fields-validation'
 export * from './email-validation'
 export * from './required-field-validation'
 export * from './validation-composite'
+export * from './validation-composite'
